api/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.3.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-appconf==1.0.5
django-cors-headers==4.2.0
django-filter==23.2
django-ipware==5.0.1
django-js-asset==2.0.0
django-kronos==1.0
django-mptt==0.14.0
django-mptt-admin==2.5.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-recaptcha==0.2.0
drf-spectacular==0.26.2
factory-boy==3.3.0
Faker==19.3.0
ghp-import==2.1.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.1
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
mergedeep==1.3.4
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.1
paginate==0.5.6
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.16.1
PyJWT==2.7.0
pymdown-extensions==10.3
pyrsistent==0.19.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
pyyaml_env_tag==0.1
redis==4.5.5
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
transliterate==1.10.2
uritemplate==4.1.1
urllib3==2.0.3
uWSGI==2.0.22
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
zeep==4.2.1
